+++ package/lib/src/beamer_back_button_dispatcher.dart
@@ -8,16 +8,22 @@ class BeamerBackButtonDispatcher extends RootBackButtonDispatcher {
   BeamerBackButtonDispatcher({
     required this.delegate,
     this.onBack,
+    this.alwaysBeamBack = false,
     this.fallbackToBeamBack = true,
   });
 
   /// A [BeamerDelegate] that belongs to the same [Router]/[Beamer] as this.
   final BeamerDelegate delegate;
 
-  /// A custom closure that has precedence over the default behavior.
+  /// A custom closure that has precedence over other behaviors.
   ///
   /// Return `true` if back action can be handled and `false` otherwise.
   final Future<bool> Function(BeamerDelegate delegate)? onBack;
+
+  /// Whether to always do [BeamerDelegate.beamBack] when Android back button
+  /// is pressed, i.e. always go to previous route in navigation history
+  /// instead of trying to pop first.
+  final bool alwaysBeamBack;
 
   /// Whether to try to use `beamBack()` when pop cannot be done.
   final bool fallbackToBeamBack;
@@ -27,6 +33,11 @@ class BeamerBackButtonDispatcher extends RootBackButtonDispatcher {
     if (onBack != null) {
       return (await onBack!(delegate));
     }
+
+    if (alwaysBeamBack) {
+      return delegate.beamBack();
+    }
+
     bool didPopRoute = await super.invokeCallback(defaultValue);
     if (didPopRoute) {
       return didPopRoute;
@@ -47,16 +58,22 @@ class BeamerChildBackButtonDispatcher extends ChildBackButtonDispatcher {
     required BeamerBackButtonDispatcher parent,
     required this.delegate,
     this.onBack,
-  })  : fallbackToBeamBack = parent.fallbackToBeamBack,
+  })  : alwaysBeamBack = parent.alwaysBeamBack,
+        fallbackToBeamBack = parent.fallbackToBeamBack,
         super(parent);
 
   /// A [BeamerDelegate] that belongs to the same [Router]/[Beamer] as this.
   final BeamerDelegate delegate;
 
-  /// A custom closure that has precedence over the default behavior.
+  /// A custom closure that has precedence over other behaviors.
   ///
   /// Return `true` if back action can be handled and `false` otherwise.
   final Future<bool> Function(BeamerDelegate delegate)? onBack;
+
+  /// Whether to always do [BeamerDelegate.beamBack] when Android back button
+  /// is pressed, i.e. always go to previous route in navigation history
+  /// instead of trying to pop first.
+  final bool alwaysBeamBack;
 
   /// Whether to try to use `beamBack()` when pop cannot be done.
   final bool fallbackToBeamBack;
@@ -66,9 +83,15 @@ class BeamerChildBackButtonDispatcher extends ChildBackButtonDispatcher {
     if (!delegate.active) {
       return false;
     }
+
     if (onBack != null) {
       return (await onBack!(delegate));
     }
+
+    if (alwaysBeamBack) {
+      return delegate.beamBack();
+    }
+
     bool didPopRoute = await super.invokeCallback(defaultValue);
     if (didPopRoute) {
       return didPopRoute;